Major Commit
Design:

Added file and preview functionality for PDF. Once users click on a capstone, the PDF file is readily accessible.
Added filters for Year and Major for easier navigation.
For Admin Sessions:

Only Home and Archive pages are visible. The "About Us" section is hidden.
Capstone Management:

Added fields for Major, Year, and Attachment File when adding a capstone.